move questions to utils
